+++ packages/bluesky/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release Note
 
+## v0.16.3
+
+- Minor improvement for `.getLabelerHeaders`. No need to aware.
+
+## v0.16.2
+
+- Add default labeler did for `.getLabelerHeaders`. No need to aware.
+
+## v0.16.1
+
+- Fixed moderation `.getLabelDefinitions` method. Now it works even if users don't have LabelersPref in their preferences.
+
 ## v0.16.0
 
 - Add `.active` and `.status` properties on `Session` and `CurrentSession` objects. ([#1516](https://github.com/myConsciousness/atproto.dart/pull/1516))