+++ packages/bluesky_chat/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release Note
 
+## v0.2.3
+
+- Bump `bluesky`.
+
+## v0.2.2
+
+- Bump `bluesky`.
+
+## v0.2.1
+
+- Bump `bluesky`.
+
 ## v0.2.0
 
 - Bump `atproto_core`, `atproto`, `bluesky`.